test(import-grouping-mismatch): externalise config tests
The inline test module grew past the 50-line
`perfectionist::excessive_inline_tests` limit; move it to
`config/tests.rs` behind an external `mod tests;`.
